update role calc, role prefs, custom role bug
- re-wrote role calculations to use new checks and transformations based
on thistleknot's latest role work
- added highlighting of which preferences are matched in roles, and
additionally listed the exact matches
- updated role preference names, and added a check to update general
material names on load
- updated roles to have proper member variables
- split the crop role preference category into crops (gather/grow) and
crops (grow)
- removed the 'wagon' creature from roles
- fixed a bug with role attributes rounding due to passing int instead
of double
- fixed a bug with roles where having invalid/bad weights resulted in
invalid ratings which caused drawing anomalies

+++ inc/attribute.h
@@ -44,16 +44,16 @@ public:
     QString get_descriptor(){return m_descriptor;}
     int get_descriptor_rank(){return m_descriptor_index;}
     int get_value() {return m_value;}
-    float get_potential_value();
-    float get_balanced_value();
+    double get_potential_value();
+    double get_balanced_value();
     void calculate_balanced_value();
     int display_value(){return m_display_value;}
-    float rating(bool potential = false);
+    double rating(bool potential = false);
     QStringList syndrome_names(){return m_syn_names;}
     float max() {return m_max;}
     float cti() {return m_cti;}
 
-    void set_rating(float rating, bool potential=false);
+    void set_rating(double rating, bool potential=false);
     void set_syn_names(QStringList names);
 
     static void load_attribute_descriptors(QSettings &s);
@@ -65,11 +65,11 @@ public:
 private:
     ATTRIBUTES_TYPE m_id;
     int m_value; //raw value including permanent syndrome effects
-    float m_value_potential;
-    float m_value_balanced;
+    double m_value_potential;
+    double m_value_balanced;
     int m_display_value; //raw value including permanent and temporary syndrome effects
     int m_max;
-    float m_rating_potential;
+    double m_rating_potential;
     float m_rating;
     int m_cti; //cost to improve (caste specific)
     QString m_descriptor; //caste specific depending on the bins
